+++ test/configuration_test.dart
@@ -174,16 +174,6 @@ Future<void> main([List<String>? args]) async {
     expect(cars.length, 1);
   });
 
-  test('Configuration readOnly - writing on read-only Realms throws', () {
-    Configuration config = Configuration.local([Car.schema]);
-    var realm = getRealm(config);
-    realm.close();
-
-    config = Configuration.local([Car.schema], isReadOnly: true);
-    realm = getRealm(config);
-    expect(() => realm.write(() {}), throws<RealmException>("Can't perform transactions on read-only Realms."));
-  });
-
   test('Configuration inMemory - no files after closing realm', () {
     Configuration config = Configuration.inMemory([Car.schema]);
     var realm = getRealm(config);